Vector store (#9)
Resolve #4

This uses a lot of dynamic dispatch, which I haven't used much before, and I think it shows. Nevertheless, code nastiness aside, it achieves the goal (reuse vectors). There are some inefficiencies (double indirections, etc.), but assuming the scripts are written for small numbers of expensive operations, which is the intended target, it should be negligible.

Note the TODOs, such as using slices instead of Vecs everywhere.